vibe-code-example/requirements.txt
scrapegraphai
python-dotenv
pydantic